requirements.txt
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-google-genai>=0.0.5 